requirements.txt
pandas==1.5.3
psycopg2-binary==2.9.4
Flask==2.2.3
Flask-RESTful==0.3.9
Werkzeug==2.2.3
python-dotenv==1.0.0
pycpfcnpj==1.7.0